+++ app/helpers/comment_deletions_helper.rb
@@ -1,0 +1,2 @@
+module CommentDeletionsHelper
+end
