+++ app/helpers/post_deletions_helper.rb
@@ -1,0 +1,2 @@
+module PostDeletionsHelper
+end
