+++ app/views/layouts/_navbar.html.erb
@@ -1,10 +1,11 @@
 <nav class="navbar navbar-expand-lg navbar-light bg-light" style="width: 100%; font-size: 14px; align-items: center;">
+
   <div class="collapse navbar-collapse" id="navbarNav">
     <ul class="navbar-nav ml-auto">
       <% if user_signed_in? %>
       <li class="nav-item" style="margin-left: 17px; margin-top: 4px;">
       <span class="navbar-text mr-3"><%= current_user.name %></span>
-    </li>
+    </li>    
         <li class="nav-item" style="padding-left: 10px;">
         <%= button_to 'Logout', destroy_user_session_path, method: :delete, class: 'btn btn-sm btn-outline-danger' %>
         </li>
@@ -15,7 +16,4 @@
       <% end %>
     </ul>
   </div>
-  <a class="navbar-brand ml-auto" href="#" style="position: absolute; top: -|4px; right: 10px;">
-    <%= image_tag 'socials-logo-noback.PNG', alt: 'Socials Logo', style: 'height: 70px;' %>
-  </a>
 </nav>
