+++ app/views/posts/index.html.erb
@@ -2,9 +2,6 @@
   <!-- Sección de usuario -->
   <h2>Users List</h2>
   <div class="d-flex align-items-start mb-4">
-    <div class="p-3">
-      <img src="path_to_default_profile_pic.jpg" alt="Profile" class="img-thumbnail" style="width: 100px; height: 100px;">
-    </div>
     <div class="border p-3 flex-grow-1 d-flex flex-column">
       <h1><%= @user.name %></h1>
       <div class="mt-auto text-right">
