+++ app/views/posts/show.html.erb
@@ -4,9 +4,6 @@
       <h2 class="text-center mb-4">Post Details</h2>
       <div class="position-relative mb-5"> <!-- This acts as the relative container, with added margin-bottom -->
         <div class="d-flex align-items-start">
-          <div class="mr-3">
-            <img src="path_to_default_profile_pic.jpg" alt="Profile" class="img-thumbnail" style="width: 100px; height: 100px;">
-          </div>
           <div class="border p-3 flex-grow-1 d-flex flex-column">
             <h1 class="text-center mb-3"><%= @post.title %></h1>
             <p class="text-justify mb-3"><%= @post.text %></p>
