--- app/views/users/sessions/new.html.erb
@@ -1,34 +1,0 @@
-<div class="vh-100 d-flex justify-content-center align-items-start">
-  <div class="w-50 mt-5">
-    <h2 class="text-center">Log in</h2>
-
-    <%= form_for(resource, as: resource_name, url: session_path(resource_name), html: { class: "mt-4" }) do |f| %>
-      <%= render "devise/shared/error_messages", resource: resource %>
-
-      <div class="form-group">
-        <%= f.label :email %>
-        <%= f.email_field :email, autofocus: true, autocomplete: "email", class: "form-control" %>
-      </div>
-
-      <div class="form-group">
-        <%= f.label :password %>
-        <%= f.password_field :password, autocomplete: "current-password", class: "form-control" %>
-      </div>
-
-      <% if devise_mapping.rememberable? %>
-        <div class="form-group form-check">
-          <%= f.check_box :remember_me, class: "form-check-input" %>
-          <%= f.label :remember_me, class: "form-check-label" %>
-        </div>
-      <% end %>
-
-      <div class="actions">
-        <%= f.submit "Log in", class: "btn btn-primary w-100" %>
-      </div>
-    <% end %>
-
-    <div class="mt-3">
-      <%= render "devise/shared/links" %>
-    </div>
-  </div>
-</div>
