+++ app/views/users/show.html.erb
@@ -1,9 +1,6 @@
 <div class="container mt-5">
   <!-- User Info Section -->
   <div class="d-flex align-items-start mb-4">
-    <div class="p-3">
-      <img src="path_to_default_profile_pic.jpg" alt="Profile" class="img-thumbnail" style="width: 100px; height: 100px;">
-    </div>
     <div class="border p-3 flex-grow-1 d-flex flex-column">
       <h4><%= @user.name %></h4>
       <p><%= @user.bio %></p>
